% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_areas.R
\name{clear_all_caches}
\alias{clear_all_caches}
\title{Clears the simple-features as well as the data cache}
\usage{
clear_all_caches()
}
\description{
Clears the simple-features as well as the data cache
}
