% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_areas.R
\name{EEZs_fishing_grounds_data}
\alias{EEZs_fishing_grounds_data}
\title{Returns all data for fishing grounds corresponding to individual EEZs}
\usage{
EEZs_fishing_grounds_data()
}
\value{
a data frame containing all data for the fishing grounds corresponding to individual EEZs
}
\description{
Returns all data for fishing grounds corresponding to individual EEZs
}
