% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FISHING_GROUNDS_ALL}
\alias{FISHING_GROUNDS_ALL}
\title{All fishing grounds data, including their WKT}
\format{
\describe{
  \item{CODE}{The alphanumeric fishing ground code}
  \item{NAME_EN}{The English name for the fishing ground}
  \item{NAME_FR}{The French name for the fishing ground}
  \item{OCEAN_AREA_SURFACE_KM2 }{The area (in km2) of the ocean part of the fishing ground}
  \item{LAT}{The latitude of the centroid for the ocean area of the fishing ground}
  \item{LON}{The longitude of the centroid for the ocean area of the fishing ground}
  \item{WKT}{The WKT for the (multi)polygon associated to the fishing ground}
}
}
\usage{
FISHING_GROUNDS_ALL
}
\description{
All fishing grounds data, including their WKT
}
\keyword{datasets}
