% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_areas.R
\name{fishing_grounds_data_by_pattern}
\alias{fishing_grounds_data_by_pattern}
\title{Returns all data (CODE, NAME_EN, NAME_FR, OCEAN_AREA_SURFACE_KM2, CENTER_LAT, CENTER_LON, WKT) for fishing ground codes matching a provided pattern}
\usage{
fishing_grounds_data_by_pattern(pattern)
}
\arguments{
\item{pattern}{The pattern (as expected by the LIKE operator) of fishing ground codes to consider (defaults to the codes for WIO and EIO)}
}
\value{
a data frame containing all data for the fishing grounds with a fishing ground code matching the provided \code{pattern}
}
\description{
Returns all data (CODE, NAME_EN, NAME_FR, OCEAN_AREA_SURFACE_KM2, CENTER_LAT, CENTER_LON, WKT) for fishing ground codes matching a provided pattern
}
\examples{
fishing_grounds_data_by_pattern("^5")
fishing_grounds_data("SYC")
}
