% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FISHING_GROUNDS_EEZ}
\alias{FISHING_GROUNDS_EEZ}
\title{Fishing grounds data for EEZs only, including their WKT}
\format{

}
\usage{
FISHING_GROUNDS_EEZ
}
\description{
Fishing grounds data for EEZs only, including their WKT
}
\keyword{datasets}
