% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_constants.R
\docType{data}
\name{IO_ALL}
\alias{IO_ALL}
\title{IOTC all IO (single area) - TO BE REDEFINED}
\format{
An object of class \code{character} of length 1.
}
\usage{
IO_ALL
}
\description{
IOTC all IO (single area) - TO BE REDEFINED
}
\keyword{datasets}
