% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_constants.R
\docType{data}
\name{IO_BOTH}
\alias{IO_BOTH}
\title{IOTC large areas (West/East) - TO BE REDEFINED}
\format{
An object of class \code{character} of length 2.
}
\usage{
IO_BOTH
}
\description{
IOTC large areas (West/East) - TO BE REDEFINED
}
\keyword{datasets}
