% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_constants.R
\docType{data}
\name{IO_F51}
\alias{IO_F51}
\title{FAO area 51 (Western IO)}
\format{
An object of class \code{character} of length 1.
}
\usage{
IO_F51
}
\description{
FAO area 51 (Western IO)
}
\keyword{datasets}
