% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_constants.R
\docType{data}
\name{IO_F5157}
\alias{IO_F5157}
\title{FAO AREA 51 and 57 combined (best representation of IOTC area)}
\format{
An object of class \code{character} of length 2.
}
\usage{
IO_F5157
}
\description{
FAO AREA 51 and 57 combined (best representation of IOTC area)
}
\keyword{datasets}
