% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_constants.R
\docType{data}
\name{IO_F57}
\alias{IO_F57}
\title{FAO area 57 (Eastern IO)}
\format{
An object of class \code{character} of length 1.
}
\usage{
IO_F57
}
\description{
FAO area 57 (Eastern IO)
}
\keyword{datasets}
