% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_constants.R
\docType{data}
\name{IO_HS}
\alias{IO_HS}
\title{IOTC high seas - TO BE REDEFINED}
\format{
An object of class \code{character} of length 1.
}
\usage{
IO_HS
}
\description{
IOTC high seas - TO BE REDEFINED
}
\keyword{datasets}
