% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_constants.R
\docType{data}
\name{IO_WEST}
\alias{IO_WEST}
\title{IOTC WIO - TO BE REDEFINED}
\format{
An object of class \code{character} of length 1.
}
\usage{
IO_WEST
}
\description{
IOTC WIO - TO BE REDEFINED
}
\keyword{datasets}
