% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_areas.R
\name{sf_for}
\alias{sf_for}
\title{Produces the spatial features for a given number of fishing grounds}
\usage{
sf_for(fishing_grounds_data, crs = "EPSG:4326", simplify = FALSE)
}
\arguments{
\item{fishing_grounds_data}{A set of fishing grounds}

\item{crs}{a coordinate reference system (defaults to EPSG:4326)}

\item{simplify}{If TRUE, the resulting spatial boundary will be simplified, otherwise it's returned as originally provided}
}
\value{
the spatial features for the provided \code{fishing_grounds_data}
}
\description{
Produces the spatial features for a given number of fishing grounds
}
