% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FISHING_GROUNDS_ALL}
\alias{FISHING_GROUNDS_ALL}
\title{All fishing grounds data, including their WKT}
\format{

}
\usage{
FISHING_GROUNDS_ALL
}
\description{
All fishing grounds data, including their WKT
}
\keyword{datasets}
