% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FISHING_GROUNDS_EEZ}
\alias{FISHING_GROUNDS_EEZ}
\title{Fishing grounds data for EEZs only, including their WKT}
\format{
\describe{
  \item{CODE}{The alphanumeric fishing ground code for the EEZ}
  \item{NAME_EN}{The English name for the EEZ fishing ground}
  \item{OCEAN_AREA_SURFACE_KM2 }{The area (in km2) of the ocean part of the EEZ}
  \item{LAT}{The latitude of the centroid for the ocean area of the EEZ}
  \item{LON}{The longitude of the centroid for the ocean area of the EEZ}
  \item{WKT}{The WKT for the (multi)polygon associated to the EEZ}
}
}
\usage{
FISHING_GROUNDS_EEZ
}
\description{
Fishing grounds data for EEZs only, including their WKT
}
\keyword{datasets}
