% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_areas.R
\name{sf_by_code}
\alias{sf_by_code}
\title{Produces the spatial features for the provided fishing ground codes}
\usage{
sf_by_code(
  fishing_ground_codes,
  crs = "EPSG:4326",
  simplify = FALSE,
  connection = DB_IOTCSTATISTICS()
)
}
\arguments{
\item{fishing_ground_codes}{A set of fishing ground codes}

\item{crs}{a coordinate reference system (defaults to EPSG:4326)}

\item{simplify}{If TRUE, the resulting spatial boundary will be simplified, otherwise it's returned as originally provided}
}
\value{
the spatial features for the fishing grounds with the provided \code{fishing_ground_codes}
}
\description{
Produces the spatial features for the provided fishing ground codes
}
