% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_gis_wkt_areas.R
\name{WKT_to_simple_feature}
\alias{WKT_to_simple_feature}
\title{Converts all rows of a dataframe produced by the \code{fishing_grounds_data}, \code{fishing_grounds_data_by_pattern} and \code{EEZs_fishing_grounds_data}
to simple fatures' instances, using the \code{WKT} column to define the spatial boundaries of the fishing ground}
\usage{
WKT_to_simple_feature(fishing_ground_data, crs = "EPSG:4326", simplify = FALSE)
}
\arguments{
\item{fishing_ground_data}{A data frame containing all information for one or more fishing grounds}

\item{crs}{a coordinate reference system (defaults to EPSG:4326)}

\item{simplify}{If TRUE, the resulting spatial boundary will be simplified, otherwise it's returned as originally provided}
}
\value{
a data frame containing all data for the fishing grounds corresponding to individual EEZs
}
\description{
Converts all rows of a dataframe produced by the \code{fishing_grounds_data}, \code{fishing_grounds_data_by_pattern} and \code{EEZs_fishing_grounds_data}
to simple fatures' instances, using the \code{WKT} column to define the spatial boundaries of the fishing ground
}
